server/pyproject.toml
[project]
name = "awos-recruitment-mcp"
version = "0.1.0"
description = "MCP server providing AI coding assistants with a discovery engine for skills, agents, and tools."
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.0",
    "fastmcp>=2.14.5",
    "posthog>=7.0.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/awos_recruitment_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
